CORE-18698 Store and retrieve filtered transactions
