+++ components/ledger/ledger-filtered-transaction-api/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.common-library'
+}
+
+description 'Corda ledger filtered transaction API'
+
+ext.cordaEnableFormatting = true
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+
+    api platform("net.corda:corda-api:$cordaApiVersion")
+    api 'org.jetbrains.kotlin:kotlin-osgi-bundle'
+    api 'net.corda:corda-ledger-common'
+    api 'net.corda:corda-crypto'
+    api project(':libs:ledger:ledger-common-data')
+}
