+++ components/ledger/ledger-filtered-transaction-api/src/main/java/net/corda/ledger/filtered/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.filtered;
+
+import org.osgi.annotation.bundle.Export;
