+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/UtxoRequestHandlerSelectorImpl.kt
@@ -7,7 +7,6 @@ import net.corda.data.ledger.persistence.FindTransactionIdsAndStatuses
 import net.corda.data.ledger.persistence.FindUnconsumedStatesByType
 import net.corda.data.ledger.persistence.LedgerPersistenceRequest
 import net.corda.data.ledger.persistence.LedgerTypes
-import net.corda.data.ledger.persistence.PersistMerkleProofIfDoesNotExist
 import net.corda.data.ledger.persistence.PersistSignedGroupParametersIfDoNotExist
 import net.corda.data.ledger.persistence.PersistTransaction
 import net.corda.data.ledger.persistence.PersistTransactionIfDoesNotExist
@@ -26,7 +25,6 @@ import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoFindSignedLed
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoFindTransactionIdsAndStatusesRequestHandler
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoFindTransactionRequestHandler
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoFindUnconsumedStatesByTypeRequestHandler
-import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoPersistMerkleProofIfDoesNotExistRequestHandler
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoPersistSignedGroupParametersIfDoNotExistRequestHandler
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoPersistTransactionIfDoesNotExistRequestHandler
 import net.corda.ledger.persistence.utxo.impl.request.handlers.UtxoPersistTransactionRequestHandler
@@ -61,6 +59,10 @@ class UtxoRequestHandlerSelectorImpl @Activate constructor(
             factoryStorage = sandbox.getSandboxSingletonService(),
             defaultContractStateVaultJsonFactory = DefaultContractStateVaultJsonFactoryImpl(),
             jsonMarshallingService = sandbox.getSandboxSingletonService(),
+            jsonValidator = sandbox.getSandboxSingletonService(),
+            merkleProofFactory = sandbox.getSandboxSingletonService(),
+            merkleTreeProvider = sandbox.getSandboxSingletonService(),
+            filteredTransactionFactory = sandbox.getSandboxSingletonService(),
             UTCClock()
         )
 
@@ -166,14 +168,6 @@ class UtxoRequestHandlerSelectorImpl @Activate constructor(
                     persistenceService,
                     externalEventResponseFactory,
                     serializationService
-                )
-            }
-            is PersistMerkleProofIfDoesNotExist -> {
-                UtxoPersistMerkleProofIfDoesNotExistRequestHandler(
-                    req,
-                    externalEventContext,
-                    externalEventResponseFactory,
-                    persistenceService
                 )
             }
             else -> {
