--- components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoPersistMerkleProofIfDoesNotExistRequestHandler.kt
@@ -1,35 +1,0 @@
-package net.corda.ledger.persistence.utxo.impl.request.handlers
-
-import net.corda.data.KeyValuePairList
-import net.corda.data.flow.event.external.ExternalEventContext
-import net.corda.data.ledger.persistence.PersistMerkleProofIfDoesNotExist
-import net.corda.data.persistence.EntityResponse
-import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
-import net.corda.ledger.persistence.common.RequestHandler
-import net.corda.ledger.persistence.utxo.UtxoPersistenceService
-import net.corda.messaging.api.records.Record
-
-class UtxoPersistMerkleProofIfDoesNotExistRequestHandler(
-    private val persistSignedGroupParametersIfDoNotExist: PersistMerkleProofIfDoesNotExist,
-    private val externalEventContext: ExternalEventContext,
-    private val externalEventResponseFactory: ExternalEventResponseFactory,
-    private val persistenceService: UtxoPersistenceService
-) : RequestHandler {
-
-    override fun execute(): List<Record<*, *>> {
-        persistenceService.persistMerkleProof(
-            persistSignedGroupParametersIfDoNotExist.transactionId,
-            persistSignedGroupParametersIfDoNotExist.groupIndex,
-            persistSignedGroupParametersIfDoNotExist.treeSize,
-            persistSignedGroupParametersIfDoNotExist.leaves,
-            persistSignedGroupParametersIfDoNotExist.hashes
-        )
-
-        return listOf(
-            externalEventResponseFactory.success(
-                externalEventContext,
-                EntityResponse(emptyList(), KeyValuePairList(emptyList()), null)
-            )
-        )
-    }
-}
