+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/persistence/LedgerPersistenceMetricOperationName.kt
@@ -8,11 +8,9 @@ enum class LedgerPersistenceMetricOperationName {
     FindTransactionWithStatus,
     FindUnconsumedStatesByType,
     FindWithNamedQuery,
-    FindMerkleProofs,
     PersistSignedGroupParametersIfDoNotExist,
     PersistTransaction,
     PersistTransactionIfDoesNotExist,
     ResolveStateRefs,
-    UpdateTransactionStatus,
-    PersistMerkleProofIfDoesNotExist
+    UpdateTransactionStatus
 }
