+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/persistence/UtxoLedgerPersistenceServiceImpl.kt
@@ -1,33 +1,25 @@
 package net.corda.ledger.utxo.flow.impl.persistence
 
 import io.micrometer.core.instrument.Timer
-import net.corda.crypto.cipher.suite.merkle.MerkleProofFactory
 import net.corda.flow.external.events.executor.ExternalEventExecutor
 import net.corda.flow.fiber.metrics.recordSuspendable
 import net.corda.ledger.common.data.transaction.SignedTransactionContainer
 import net.corda.ledger.common.data.transaction.TransactionStatus
 import net.corda.ledger.common.data.transaction.TransactionStatus.Companion.toTransactionStatus
-import net.corda.ledger.utxo.data.transaction.MerkleProofDto
 import net.corda.ledger.utxo.data.transaction.SignedLedgerTransactionContainer
 import net.corda.ledger.utxo.flow.impl.cache.StateAndRefCache
-import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.FindMerkleProofs
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.FindSignedLedgerTransactionWithStatus
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.FindTransactionIdsAndStatuses
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.FindTransactionWithStatus
-import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.PersistMerkleProofIfDoesNotExist
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.PersistTransaction
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.PersistTransactionIfDoesNotExist
 import net.corda.ledger.utxo.flow.impl.persistence.LedgerPersistenceMetricOperationName.UpdateTransactionStatus
-import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindMerkleProofsExternalEventFactory
-import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindMerkleProofsParameters
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindSignedLedgerTransactionExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindSignedLedgerTransactionParameters
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindTransactionExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindTransactionIdsAndStatusesExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindTransactionIdsAndStatusesParameters
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.FindTransactionParameters
-import net.corda.ledger.utxo.flow.impl.persistence.external.events.PersistMerkleProofIfDoesNotExistExternalEventFactory
-import net.corda.ledger.utxo.flow.impl.persistence.external.events.PersistMerkleProofIfDoesNotExistParameters
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.PersistTransactionExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.PersistTransactionIfDoesNotExistExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.persistence.external.events.PersistTransactionIfDoesNotExistParameters
@@ -47,7 +39,6 @@ import net.corda.utilities.serialization.deserialize
 import net.corda.v5.application.serialization.SerializationService
 import net.corda.v5.base.annotations.Suspendable
 import net.corda.v5.crypto.SecureHash
-import net.corda.v5.crypto.merkle.MerkleProof
 import net.corda.v5.ledger.common.transaction.CordaPackageSummary
 import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction
 import net.corda.v5.serialization.SingletonSerializeAsToken
@@ -75,9 +66,7 @@ class UtxoLedgerPersistenceServiceImpl @Activate constructor(
     @Reference(service = UtxoSignedTransactionFactory::class)
     private val utxoSignedTransactionFactory: UtxoSignedTransactionFactory,
     @Reference(service = StateAndRefCache::class)
-    private val stateAndRefCache: StateAndRefCache,
-    @Reference(service = MerkleProofFactory::class)
-    private val merkleProofFactory: MerkleProofFactory
+    private val stateAndRefCache: StateAndRefCache
 ) : UtxoLedgerPersistenceService, UsedByFlow, SingletonSerializeAsToken {
 
     @Suspendable
@@ -163,31 +152,6 @@ class UtxoLedgerPersistenceServiceImpl @Activate constructor(
     }
 
     @Suspendable
-    override fun findMerkleProofs(
-        transactionId: SecureHash,
-        groupIndex: Int
-    ): List<MerkleProof> {
-        return recordSuspendable({ ledgerPersistenceFlowTimer(FindMerkleProofs) }) @Suspendable {
-            wrapWithPersistenceException {
-                externalEventExecutor.execute(
-                    FindMerkleProofsExternalEventFactory::class.java,
-                    FindMerkleProofsParameters(transactionId.toString(), groupIndex)
-                )
-            }.firstOrNull()?.let {
-                serializationService.deserialize<List<MerkleProofDto>>(it.array()).map { merkleProofDto ->
-                    merkleProofFactory.createAuditMerkleProof(
-                        merkleProofDto.transactionId,
-                        merkleProofDto.groupIndex,
-                        merkleProofDto.treeSize,
-                        merkleProofDto.leavesWithData,
-                        merkleProofDto.hashes
-                    )
-                }
-            } ?: emptyList()
-        }
-    }
-
-    @Suspendable
     override fun persist(
         transaction: UtxoSignedTransaction,
         transactionStatus: TransactionStatus,
@@ -234,22 +198,6 @@ class UtxoLedgerPersistenceServiceImpl @Activate constructor(
                     "V" -> TransactionExistenceStatus.VERIFIED
                     else -> throw IllegalStateException("Invalid status $status")
                 } to summaries
-            }
-        }
-    }
-
-    @Suspendable
-    override fun persistMerkleProofIfDoesNotExist(
-        transactionId: SecureHash,
-        groupIndex: Int,
-        merkleProof: MerkleProof
-    ) {
-        recordSuspendable({ ledgerPersistenceFlowTimer(PersistMerkleProofIfDoesNotExist) }) @Suspendable {
-            wrapWithPersistenceException {
-                externalEventExecutor.execute(
-                    PersistMerkleProofIfDoesNotExistExternalEventFactory::class.java,
-                    PersistMerkleProofIfDoesNotExistParameters(transactionId.toString(), groupIndex, merkleProof)
-                )
             }
         }
     }
