+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/filtered/UtxoFilteredTransactionInternal.kt
@@ -1,0 +1,9 @@
+package net.corda.ledger.utxo.flow.impl.transaction.filtered
+
+import net.corda.ledger.filtered.FilteredTransaction
+import net.corda.v5.ledger.utxo.transaction.filtered.UtxoFilteredTransaction
+
+interface UtxoFilteredTransactionInternal : UtxoFilteredTransaction {
+
+    val filteredTransaction: FilteredTransaction
+}
