+++ gradle.properties
@@ -42,7 +42,7 @@ commonsLangVersion = 3.12.0
 commonsTextVersion = 1.10.0
 # Corda API libs revision (change in 4th digit indicates a breaking change)
 # Change to 5.2.0.xx-SNAPSHOT to pick up maven local published copy
-cordaApiVersion=5.2.0.29-beta+
+cordaApiVersion=5.2.0.29-SNAPSHOT
 
 disruptorVersion=3.4.4
 felixConfigAdminVersion=1.9.26
