+++ libs/crypto/cipher-suite/build.gradle
@@ -14,6 +14,7 @@ dependencies {
     implementation project(":libs:crypto:crypto-utils")
 
     api platform("net.corda:corda-api:$cordaApiVersion")
+    api 'net.corda:corda-application'
     api 'net.corda:corda-crypto'
     api 'net.corda:corda-crypto-extensions'
