+++ libs/crypto/merkle-impl/src/main/kotlin/net/corda/crypto/merkle/impl/MerkleProofFactoryImpl.kt
@@ -1,7 +1,9 @@
 package net.corda.crypto.merkle.impl
 
 import net.corda.crypto.cipher.suite.merkle.MerkleProofFactory
+import net.corda.sandbox.type.SandboxConstants.CORDA_MARKER_ONLY_SERVICE
 import net.corda.sandbox.type.UsedByFlow
+import net.corda.sandbox.type.UsedByPersistence
 import net.corda.v5.base.annotations.Suspendable
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.crypto.merkle.MerkleProof
@@ -12,11 +14,12 @@ import org.osgi.service.component.annotations.Component
 import org.osgi.service.component.annotations.ServiceScope
 
 @Component(
-    service = [ MerkleProofFactory::class, UsedByFlow::class ],
+    service = [ MerkleProofFactory::class, UsedByFlow::class, UsedByPersistence::class ],
+    property = [ CORDA_MARKER_ONLY_SERVICE ],
     scope = ServiceScope.PROTOTYPE
 )
 class MerkleProofFactoryImpl @Activate constructor()
-    : MerkleProofFactory, UsedByFlow, SingletonSerializeAsToken {
+    : MerkleProofFactory, UsedByFlow, UsedByPersistence, SingletonSerializeAsToken {
 
     @Suspendable
     override fun createAuditMerkleProof(
