+++ libs/crypto/merkle-impl/src/main/kotlin/net/corda/crypto/merkle/impl/MerkleProofImpl.kt
@@ -1,5 +1,6 @@
 package net.corda.crypto.merkle.impl
 
+import net.corda.crypto.cipher.suite.merkle.MerkleProofInternal
 import net.corda.v5.base.exceptions.CordaRuntimeException
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.crypto.extensions.merkle.MerkleTreeHashDigestProvider
@@ -28,7 +29,7 @@ class MerkleProofImpl(
     private val treeSize: Int,
     private val leaves: List<IndexedMerkleLeaf>,
     private val hashes: List<SecureHash>
-) : MerkleProof {
+) : MerkleProofInternal {
     // CORE-5111: add serialize/deserialize (and its test)
 
     override fun verify(root: SecureHash, digest: MerkleTreeHashDigest) =
@@ -304,7 +305,7 @@ class MerkleProofImpl(
      * @return A MerkelProofImpl which has the union of the leaves and the required proof hashes to be verifiable.
      */
     @Suppress("UnusedParameters")
-    fun merge(other: MerkleProofImpl, digest: MerkleTreeHashDigestProvider): MerkleProofImpl {
+    override fun merge(other: MerkleProof, digest: MerkleTreeHashDigestProvider): MerkleProof {
         require(this.treeSize == other.treeSize) {
             "underlying tree sizes must match; left hand side has underlying tree "+
                 "size ${this.treeSize} and right hand side has underlying tree size ${other.treeSize}"
@@ -329,7 +330,7 @@ class MerkleProofImpl(
             nodeMapThis[k] = it
         }
         val nodeMapOther: MutableMap< Pair<Int, Int>, MerkleNodeInfo> = mutableMapOf()
-        val otherRoot = other.calculateRootInstrumented(digest) {
+        val otherRoot = (other as MerkleProofImpl).calculateRootInstrumented(digest) {
             val k = it.level to it.node.indexWithinLevel
             nodeMapOther[k] = it
         }
