+++ libs/ledger/ledger-utxo-data/src/main/kotlin/net/corda/ledger/utxo/data/transaction/UtxoFilteredTransactionDto.kt
@@ -1,0 +1,12 @@
+package net.corda.ledger.utxo.data.transaction
+
+import net.corda.ledger.common.data.transaction.PrivacySalt
+import net.corda.v5.base.annotations.CordaSerializable
+
+@CordaSerializable
+data class UtxoFilteredTransactionDto(
+    val transactionId: String,
+    val merkleProofMap: Map<Int, List<MerkleProofDto>>,
+    val privacySalt: PrivacySalt,
+    val metadataBytes: ByteArray
+)
