+++ settings.gradle
@@ -130,6 +130,8 @@ include 'components:kafka-topic-admin'
 include 'components:ledger:ledger-common-flow'
 include 'components:ledger:ledger-common-flow-api'
 include 'components:ledger:ledger-consensual-flow'
+include 'components:ledger:ledger-filtered-transaction-api'
+include 'components:ledger:ledger-filtered-transaction'
 include 'components:ledger:notary-worker-selection'
 include 'components:ledger:notary-worker-selection-impl'
 include 'components:ledger:ledger-persistence:testing-datamodel'
